ref: Move path builders to separate file
