+++ src/utils/dateUtils.ts
@@ -437,3 +437,22 @@ export function isValidThesisDate(dateString: string): boolean {
 export function getCurrentThesisDate(): string {
     return formatThesisDate(new Date());
 }
+
+/**
+ * Get the current academic year in the format "YYYY-YYYY" (e.g., "2025-2026").
+ * Academic year typically starts in June/August, so:
+ * - Jan-May: previous year to current year (e.g., 2024-2025)
+ * - Jun-Dec: current year to next year (e.g., 2025-2026)
+ * @param date - Optional date to calculate academic year for (defaults to current date)
+ * @returns Academic year string in format "YYYY-YYYY"
+ */
+export function getAcademicYear(date: Date = new Date()): string {
+    const year = date.getFullYear();
+    const month = date.getMonth(); // 0-indexed (0 = January)
+    // If before June (month < 5), we're in the second half of the academic year
+    if (month < 5) {
+        return `${year - 1}-${year}`;
+    }
+    // June onwards, we're in the first half of a new academic year
+    return `${year}-${year + 1}`;
+}
